chore: move auto-refresh test to OSS
